Allow AD to use primitive adjoints that provide a superset of indices (#19751)
* Mark primitive vs. synthetic primal/adjoint in [reverse-differentiable].

* Allow synthesis of gradient functions using a subset of adjoint indices.

Each gradient instruction now creates a canonical (and potentially 
non-canonical) gradient function with the requested set of indices. This 
function will call the primal and adjoint with the minimal index 
superset if one exists, or synthesize a primal/adjoint with exactly the 
requested set if not. Also adds a test for this behavior.

* Only enable superset behavior for primitively specified adjoints.

Adjoint selection during gradient and nested adjoint synthesis now
prefers the first available option of a) using an existing adjoint with
exactly the right indices, b) using the primitively specified adjoint
with the smallest available superset of indices, or c) synthesizing a
new adjoint with the exact indices.

* Address comments

* add broken tests

* Switch from !synthesized to primitive and fix logic and tests.

Marking primitive rather than synthesized makes the state of a newly 
created [reverse_differentiable] attribute (which doesn't have an 
adjoint yet but should still have had synthesized marked) less 
confusing.