Force AttributeBase Alignment to 8-byte
On 32-bit ARM, if an AvailableAttr was getting 4-byte aligned, it would trigger a LLVM assert that the PointerIntPair wasn't aligned enough. Forcing 8-byte alignment for AttributeBase addresses the problem. However, the assert is mentioning it wants 2 bits, so I'm not 100% sure where the extra bit is coming from.

This is a little messy, since if you are using the custom new operator on a child class of AttributeBase, you get the implementation on AttributeBase, which forces alignment to be that of AttributeBase. This is problematic if any child class requires more strict alignment than AttributeBase in general.