Sema: Add TypeResolution::resolveSelfAssociatedType()
We used to have a similar method on GenericTypeResolver for resolving
associated types found by unqualified lookup.

Bring it back so that we can handle unqualified and qualified (with a
'Self.' prefix) lookups in the same code path, which fixes some issues
with ambiguities.

For now, the new code path is mostly disabled, because it interacts
badly with associated type inference. But once that is fixed, more
cleanup will be possible here.