[android] Disable LLD when cross-compiling Android in CI.
Ubuntu 16.04 doesn't include LLD, and the included gold doesn't
link Android correctly.

Because of the change in e7ba154d59d73c5bfa04090c0d32f187cc6ad873
the -B option was not passed during compiling, and the Ubuntu gold
was being used in the CI (the default value for
SWIFT_ENABLE_LLD_LINKER is ON, even if LLD is not present).

This change sets the value to OFF in the preset, so LLD will not
try to be used in CI. All other builds should be unchanged, so in
other developers setups the compiling machine LLD will try to be
used (which works on more recent Ubuntu versions).