build: split up and rename `add_swift_library`
Split up the `add_swift_library` by extracting the tools host handling
aspect of it.  This has a much more limited option set.  Reducing the
options further would allow us to eliminate this and replace it with a
wrapper around `add_llvm_library`.  Leave the bulk of the logic in place
in the renamed `swift_add_target_library` which can now be simplified on
its own.