[android] Adapt POSIX test to Android
Android differs in subtle ways from Linux. For example sem_open is
completely unimplemented, and always returns ENOSYS. None of the
test using sem_open were going to work.

Additionally the l_len field of the flock struct is typed as
__kernel_off_t, which doesn't have the same size as off_t.

Finally, the current working directory is always /, so creating
relative files will fail. We have to find the executable directory
and create the file there.