[ConstraintSystem] Allow getPotentialBindings to find supertypes of optionals.
Improve enumerateDirectSupertypes so that for T? it will return U? if
U is a supertype of T. This is another form of direct supertype.

This is making up for a deficiency in the completeness of our Type
join implementation, which should be able to directly compute a join
of disparate types that share some common supertype, but sometimes
fails to in cases involving protocol compositions.

Fixes rdar://problem/45490737

(cherry picked from commit 85237aa0717d9ffc709326a4f7de41e8f717cf34)