IRGen: inline the single use macro
Inline the single use macro for the sizeof the header.  This allows the compiler
to build with Visual Studio which does not support C11.  Avoiding the header
avoids the use of SwiftStddef.h which is meant for the runtime/stdlib only and
uses C11 assuming that it will only be built with clang.