[SE-0155] Default Arguments in Enum Cases
Lift the restriction on default arguments in enum cases.  This
necessitated a number of odd tradeoffs stemming mostly from the
fact that enum cases are not a form of decl context.

An entity has been added to the mangling for default arguments that
mangle the enum case in addition to the entity's decl context.

SILGenExpr has regrown the ability to emit tuple shuffles that involve
caller and callee-side default arguments to support the open-coding path
in SILGenApply.  This could probably be made to use the machinery there
to cough up an RValue that we could use as the payload, but that
involves some deep refactoring.