[build-preset] Add no_test variants of all of the buildbot_linux_* presets.
The key thing is that these skip all testing except for the integration testing
and thus are perfect for subproject buildbot testing.

The reason why I am doing this is I discovered that dispatch is not using
build-presets. This means that dispatch may not get updated properly (resulting
in integration test failures) when we add new package installation dependencies
(like llvm-cov, llvm-profdata) for the integration tests.

This is functionally the same as dispatch's build today except that we /will/
build lldb. That being said I think it makes sense to build the entire package
for dispatch so if we have any integration issues in between lldb and dispatch
we can test them.

SR-9080