Sema: Fix another crash with an invalid extension
Perhaps we should refactor everything so that extensions
that are not at the top-level are still bound and type
checked normally. However that requires a bit of work, so
keep playing wack-a-mole for now to handle these invalid
states when they come up.

However while we're at it, make the code a little better
by removing a bogus diagnostic path that was not used.

Fixes <rdar://problem/45290211>, <https://bugs.swift.org/browse/SR-9009>.