[android] Do  not link library absolute path.
Avoid using the full path of the shared library to avoid including
the absolute path of the library in the executable, which will not
work once the library is moved to the host machine (like an
Android device).

To avoid doing a confussing `-lrary`, I changed the name of the
library to `libAttrImplFP.so` instead.