Fixed a SILVerifier failure, where a synthesized host inst for switch_case based (#20141)
control flow handling is a ReturnKind, because its SILLocation came from the
first instruction of the BasicBlock to which we are adding host code for
"sending case ID to tensorflow".

When that instruction came from a return statement originally, its SILLocation
is ReturnKind. The fix is to overwrite that location kind to be RegularKind.

We will also start a thread with Swift core team / community on the background of
LocationKind, and whether the fix can be made simpler.