Add a bunch more 'try' tests.
Some of these are currently failing. For example:

```swift
// expected-error {{cannot convert value of type 'Int??' to specified type 'Int'}}
let _: Int = try? producer.produceDoubleOptionalInt()
```

This actually produces this warning:

```
value of optional type 'Int?' not unwrapped; did you mean to use 'try!' or chain with '?'?
```

Note that the value in question is not an `Int?`, so a suggestion of `try!` here is inappropriate.
(And, due to the change in semantics of 'try?', wouldn't change anything anyway.)